Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import blogbench-1.0

Blogbench is a portable filesystem benchmark that tries to reproduce the
load of a real-world busy file server.
It stresses the filesystem with multiple threads performing random
reads, writes and rewrites in order to get a realistic idea of the
scalability and the concurrency a system can handle.

ok bernd@