Written by: Matthias Kilian <kili@openbsd.org>
Since their world was invaded by an alien race, the Blobs have faced
a lifetime of war. But now they have a chance to win the war once
and for all.

In Blob Wars : Metal Blob Solid, you take on the role of a fearless
Blob agent, Bob. Bob's mission is to infiltrate the various enemy
bases around the Blobs' homeworld and rescue as many MIAs as possible.
But standing in his way are many vicious aliens, other Blobs who
have been assimilated and the evil alien leader, Galdov.

import! import! ajacoutot@