Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import subtitleeditor-0.13.6

Subtitle Editor is a GTK+2 tool to edit subtitles.
It can be used for new subtitles or as a tool to transform, edit,
correct and refine existing subtitle. This program also shows sound
waves, which makes it easier to synchronise subtitles to voices.

Supported formats are:
 * Sub Station Alpha
 * Advanced Sub Station Alpha
 * SubRip
 * MicroDVD
 * MPL2
 * MPsub (MPlayer subtitle)
 * SubViewer 2.0
 * Plain-Text
 * Adobe Encore DVD
 
ok steven@