Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Start importing some pieces of the GNUstep framework.
GNUstep is a cross-platform, object-oriented framework for desktop
application development. Based on the OpenStep specification originally
created by NeXT (now Apple), GNUstep enables developers to rapidly build
sophisticated software by employing a large library of reusable software
components.

Not connected to the build yet as this is a work in progress and there
might be some glitches.
Now that it's in CVS, hopefully more people will test this.

ok robert@