Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import gnustep-make-2.0.1

The makefile package is a simple, powerful and extensible way to write
makefiles for a GNUstep-based project.  It allows the user to write a
project without having to deal with the complex issues associated with
configuration, building, installation, and packaging.  It also allows
the user to easily create cross-compiled binaries.