Written by: Uwe Stuehler <uwe@openbsd.org>
Update devel/bluetooth-libs and net/bluetooth-tools to match the kernel

This time, the userland tools and libraries are being ported from NetBSD
(which in turn is also based on FreeBSD code).  Both packages represent
work in progress and are not fully functional yet.

ok jasper@ for the new ports, grange@ for taking over maintainership