Written by: Martynas Venckus <martynas@openbsd.org>
import py-kid-0.9.5

Kid is a simple template language for XML based vocabularies written
in Python. It was spawned as a result of a kinky love triangle
between XSLT, TAL, and PHP. We believe many of the best features
of these languages live on in Kid with much of the limitations and
complexity stamped out (see WhatsBorrowed and WhatsDifferent).

ok simon@