Written by: Matthias Kilian <kili@openbsd.org>
ISO Master is an open-source, easy to use, GUI CD image editor.
Basically you can use this program to extract files from an ISO, 
add files to an ISO, and create bootable ISOs.
It can open both ISO and NRG files but can only save as ISO.

From Giovanni Bechis <bigionews () snb ! it>

(And this time in the correct directory)