Written by: Reyk Floeter <reyk@openbsd.org>
Add security/wpa_supplicant, a fully-featured implementation of an
IEEE 802.1X supplicant. This port does not support the wireless
WPA/WPA functionality yet, because it is missing in the OpenBSD
net80211 kernel implementation. Successfully tested with 802.1X on
ProCurve switches with OpenBSD running freeradius in the backend.

This port is based on work by Jussi Salzwedel, thanks!

ok aanriot@