Written by: Simon Bertrang <simon@openbsd.org>
import vodovod-1.10

The goal of vodovod is to reach the highest possible score. You get a
limited number of pipes on each level and need to combine them to lead
the water from the house at the top of the screen to the storage tank at
the bottom.

help & ok jasper@