Written by: Marc Espie <espie@openbsd.org>
take over amule as MAINTAINER, replaces mldonkey nicely for me
Activate a lot of useful stuff, like the amule daemon and the amulecmd
line tool, which makes it possible to run amule mostly unattended.
(you might need to run it with the gui to configure it at first).

Even the experimental amulegui appears to be mostly functional.
the webgui is included for reference, it works on i386, something fishy
on sparc64...

todo: figure out cross-platform issues. The people who write amule never
thought that the client and the server could run on distinct architectures,
and so they transfer binary data without even putting it in network format,
leading to interesting transfer times and sizes...