Written by: Simon Bertrang <simon@openbsd.org>
import mod_scgi

This is an Apache module that implements the SCGI protocol.  The SCGI
protocol is a replacement for the Common Gateway Interface (CGI)
protocol. It is a standard for applications to interface with HTTP
servers. It is similar to FastCGI but is designed to be easier to
implement.

from Jeremy Evans, with tweaks by me