Written by: Stuart Henderson <sthen@openbsd.org>
fix pear build with safe_mode=on in php.ini by using -n to ignore
the ini file. package version not bumped as no plist_db conflicts or
package change (it just failed to build if safe mode was enabled in
/var/www/conf/php.ini).

checked pear modules and stuff.

"please build all the pear modules and stuff if it works
ok robert@"