Written by: Michael Erdely <merdely@openbsd.org>
mod_fastcgi is a module for the Apache2 web server, that enables
FastCGI - a standards based protocol for communicating with
applications that generate dynamic content for web pages.

FastCGI provides a superset of CGI functionality, but a subset of the
functionality of programming for a particular web server API.
Nonetheless, the feature set is rich enough for programming virtually
any type of web application, but the result is generally more
scalable.

Original suggestion from Roland Philippsen
Advice from steve@ and mbalmer@
Help and ok, simon@