Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import belooted-0.1.3

Belote is a popular 32-card trick-taking game played in France.
It derived around 1920, probably via Clobyosh, from Klaverjassen, a game
played since at least the 1600s in the Netherlands. Closely related
games are played throughout the world.
Within the game's terminology, belote is used to designate a pair of a
King and a Queen of a trump suit, possibly yielding the game's name
itself.


ok simon@