Written by: Alexandre Anriot <aanriot@openbsd.org>
enable the "X-Forwarded-For" headers support which has been added in
Squid 2.6. It's not the default in squid.conf but can be enabled, and
it's useful e.g. when you are using squid through dansguardian, in order
to keep/show the source address of the client in the logs (otherwise,
squid thinks that everything comes from dansguardian).

ok brad@