Written by: Simon Bertrang <simon@openbsd.org>
import p5-Chart-Math-Axis

Chart::Math::Axis implements in a generic way an algorithm for finding a
set of ideal values for an axis.  That is, for any given set of data,
what should the top and bottom of the axis scale be, and what should the
interval between the ticks be.

ok jasper@