Written by: Simon Bertrang <simon@openbsd.org>
import p5-Validate-Net

Validate::Net is a class designed to assist with the validation of
internet related strings.  It can be used to validate CGI forms,
internally by modules, and in any place where you want to check that an
internet related string is valid before handing it off to Net::*
modules.

ok jasper@