Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import libgalago 0.5.2

Galago is a desktop presence framework, designed to transmit presence
information between programs. To put it in simpler terms, it takes
information on who is online and their away/idle states from an instant
messenger (such as gaim) or other similar programs and lets other
programs (such as Ximian's Evolution) to make use of it.

The advantage of such a framework is that it brings your programs closer
together. When you receive an e-mail from a friend who is in your buddy
list, you'll be able to immediately tell their online status, for
example.

feedback and ok ajactoutot@