Written by: Theo de Raadt <deraadt@openbsd.org>
on further inspection of the terms UW wants to hold us to, I have
decided that these legal assholes don't even deserve to be on our ftp
site.  Yes, this is probably the first time I have ever commited to
the ports tree, but this is important.  When people like that come and
fool our developers, I will fight back against them.  pine is not free
in any sense, and we will no longer distribute it.