Written by: Stuart Henderson <sthen@openbsd.org>
import p5-File-BOM

This module provides functions for handling unicode byte order
marks, which are to be found at the beginning of some files and
streams.

The intention of File::BOM is for files with BOMs to be readable
as seamlessly as possible, regardless of the encoding used.

ok simon@