Written by: Kurt Miller <kurt@openbsd.org>
Exit prompt loop when there's nothing to read from stdin. This can happen
when stdin is redirected someplace unanticipated. This fixes the infinite
prompt loop when the port is built as a build depend of another port
(reported by Martin Toft <mt at martintoft.dk>). Still to do is fix
bsd.port.mk to not redirect stdin for build depends...