Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import alpine-0.99999

Alpine is a screen-oriented message-handling tool.  In its default
configuration, Alpine offers an intentionally limited set of functions
geared toward the novice user, but it also has a large list of optional
"power-user" and personal-preference features.


Note: Alpine is the next-generation successor to Pine and supersed it;
it is not a rewrite, it is mostly the same code with a license change
and several improvements (Pine development stopped more than 2 years ago).

ok djm@