Written by: Matthias Kilian <kili@openbsd.org>
Update to 2.20.2 (lots of bugfixes and required due to the poppler
update).

Based on an update to 0.9.* from Mikolaj Kucharski several months ago.

Note that this doesn not include the no_libgnome FLAVOR suggested by
Mikolaj, because it saves much less dependencies and space than it did
for the 0.9.* version.

ok landry@