Written by: Michael Erdely <merdely@openbsd.org>
Initial import of py-clientform-0.2.7

ClientForm is a Python module for handling HTML forms on the client
side, useful for parsing HTML forms, filling them in and returning the
completed forms to the server. It developed from a port of Gisle Aas'
Perl module HTML::Form, from the libwww-perl library, but the interface
is not the same.

From Will Maier
ok eric@