Written by: Marc Balmer <mbalmer@openbsd.org>
The Horde Team is pleased to announce the final release of the Kronolith
Calendar Application version H3 (2.1.7).

This is a bugfix release that also fixes privilege escalation in the Horde API
and missing ownership validation in the share management. All users are
encouraged to upgrade to this version.

Major changes compared to the Kronolith H3 (2.1.6) version are:
    * Fixed privilege escalation in the Horde API.
    * Fixed missing ownership validation on share changes.
    * Updated Japanese translation.
    * Small bugfixes and improvements.

The full list of changes (from version H3 (2.1.6)) can be viewed here:
http://cvs.horde.org/diff.php/kronolith/docs/CHANGES?r1=1.165.2.158&r2=1.165.2.161.2.3&ty=h