Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import gns3-0.3.2

GNS3 is a graphical network simulator that allows you to design complex
network topologies and to launch simulations on them.
It is an excellent complementary tool to real labs for administrators of
Cisco networks or people wanting to pass their CCNA, CCNP, CCIP or CCIE
certifications.  It can also be used to experiment features of Cisco IOS
or to check configurations that need to be deployed later on real
routers.


Based on an original submission by MAINTAINER Jonathan Armani <dbd at asystant dot net>