Written by: Bernd Ahlers <bernd@openbsd.org>
Initial import of ruby-mocha-0.5.6.

Mocha is a library for mocking and stubbing using a syntax like that of
JMock, and SchMock. Most commonly Mocha is used in conjunction with
Test::Unit, but it can be used in other contexts.

One of its main advantages is that it allows you to mock and stub methods
on real (non-mock) classes and instances. You can for example stub
ActiveRecord instance methods like create, save, destroy and even class
methods like find to avoid hitting the database in unit tests.

Mocha provides a unified, simple and readable syntax for both traditional
mocking and for mocking with real objects.