Written by: Landry Breuil <landry@openbsd.org>
Import xfprint, based on an initial submission somes time ago by vlad
glagolev. It is a graphical frontend for printing and printer management,
supporting CUPS, BSD-LPR and File(Postscript) backends.
Add a patch from svn to fix a bug when no printer was selected as default.

Feedback and ok ajacoutot@