Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import krename-3.0.14

KRename is a powerful batch renamer for KDE.  It allows you to easily
rename hundreds or even more files in one go.  The filenames can be
created by parts of the original filename, numbering the files or
accessing hundreds of informations about the file, like creation date or
Exif informations of an image, thanks to KDE file plugins.


Based on a submission by Vadim Zhukov
<persgray at gmail dot com> (MAINTAINER)